requirements.txt
fastapi
uvicorn
langchain
langchain_community
langchain_openai
beautifulsoup4
python-dotenv